增加双人对战版

+++ src/handles/roomHandle.js
@@ -10,26 +10,35 @@ function getRoomInfoByType(type) {
         }
         const user_id = roomService.getUserId() + "000"
         const room_id = roomService.getRoomId() + "000"
-        return roomService.putRunningRoomInfo(room_id, user_id)
+        yield roomService.putRunningRoomInfo(room_id, user_id)
+        return [room_id, user_id]
     })
 
 }
 
 module.exports = (io, socket) => {
+
+    if (socket.request.session.$user) {
+        const { room_id, user_id, role } = socket.request.session.$user
+        socket.join(room_id)
+    }
     socket.on(event.wuzi.init, (data) => {
         co(function*() {
             const { type } = data
-            if (socket.request.$user) { //恢复
-                const { room_id, user_id } = socket.request.$user
-                const chessList = yield roomService.getChessList()
-                const [belong_user_id, custom_user_id] = roomService.getRunningRoomInfo(room_id)
+
+            if (socket.request.session.$user) { //恢复
+                const { room_id, user_id, role } = socket.request.session.$user
+                const chessList = yield roomService.getChessList(room_id, role)
+                const [belong_user_id, custom_user_id] = yield roomService.getRunningRoomInfo(room_id)
                 const other_user_id = belong_user_id === user_id ? custom_user_id : belong_user_id
-                if (chessList && belong_user_id && custom_user_id) {
+                if (chessList.length > 0 && belong_user_id && custom_user_id) {
                     if (type === "computer") {
                         const aiService = new AiService(room_id)
-                        aiService.restore(chessList)
+                        aiService.restore(chessList, role)
                     }
-                    return socket.emit(event.wuzi.restore, { status: "restore", data: { user_id, other_user_id, chessList } })
+                    socket.join(room_id)
+                    const should_play_role = yield roomService.getShouldPlayRole(room_id, user_id)
+                    return socket.emit(event.wuzi.restore, { status: "restore", data: { user_id, other_user_id, chessList, player: role, should_play_role } })
                 }
             }
 
@@ -39,7 +48,7 @@ module.exports = (io, socket) => {
             const room_info = yield getRoomInfoByType(type)
             const room_id = room_info ? room_info[0] : make_room_id
             if (!room_info) {
-                yield roomService.putAloneRoomInfo(make_room_id, socket_id)
+                yield roomService.putAloneRoomInfo(make_room_id, user_id)
             }
             const status = room_info ? "start" : "waiting"
             const other_user_id = room_info ? room_info[1] : null
@@ -53,13 +62,15 @@ module.exports = (io, socket) => {
                 socket.leave(room_id)
             })
             socket.join(room_id)
-            socket.emit(event.wuzi.init, { status, data: { user_id, other_user_id } })
-            socket.request.$user = {
-                room_id: make_room_id,
+            const should_play_role = yield roomService.getShouldPlayRole(room_id, user_id)
+            socket.emit(event.wuzi.init, { role, status, should_play_role })
+            socket.request.session.$user = {
+                room_id,
                 user_id,
                 role,
                 type
             }
+            socket.request.session.save()
             if (room_info) {
                 socket.to(room_id).emit(event.wuzi.other_come, { other_user_id: user_id })
             }
@@ -74,16 +85,16 @@ module.exports = (io, socket) => {
 
     socket.on(event.wuzi.go, (data) => {
         co(function*() {
-            const { room_id, role, type } = socket.request.$user
+            const { room_id, role, type } = socket.request.session.$user
             const { x, y } = data
             yield roomService.putChess(room_id, x, y, role)
             if (type == "computer") {
                 const aiService = new AiService(room_id)
                 const p = aiService.go(x, y)
-                yield roomService.putChess(room_id, p.x, p.y, !role)
+                yield roomService.putChess(room_id, p[0], p[1], Number(!role))
                 return socket.emit(event.wuzi.go, p)
             }
-            return socket.to(room_id).emit(event.wuzi.go, data)
+            return socket.to(room_id).emit(event.wuzi.go, [x, y])
         })
 
 
@@ -91,22 +102,33 @@ module.exports = (io, socket) => {
 
     socket.on(event.wuzi.back, (data) => {
         co(function*() {
-            const { room_id, role, type } = socket.request.$user
+            const { room_id, role, type } = socket.request.session.$user
             const { x, y } = data
             yield roomService.backChess(room_id, x, y, role)
             if (type == "computer") {
                 const aiService = new AiService(room_id)
                 const p = aiService.back()
-                yield roomService.backChess(room_id, p.x, p.y, !role)
+                yield roomService.backChess(room_id, p[0], p[1], Number(!role))
             }
             socket.to(room_id).emit(event.wuzi.back, data)
         })
     })
 
-    socket.on(event.wuzi.finish, function(data) {
-        const { room_id, role, type } = socket.request.$user
-        yield roomService.removeChessList(room_id)
-        yield roomService.removeRunningRoom(room_id)
+    socket.on(event.wuzi.finish, (data) => {
+        co(function*() {
+            if (socket.request.session.$user) {
+                const { room_id, role, type } = socket.request.session.$user
+                yield roomService.removeChessList(room_id)
+                yield roomService.removeRunningRoom(room_id)
+                if (type == "computer") {
+                    const aiService = new AiService(room_id)
+                    aiService.destroy()
+                }
+            }
+
+            socket.request.session.$user = {}
+            socket.request.session.save()
+        })
     })
 
 }
